Move InputStreamHelper to api
